Conflicts:
	rec-web/src/main/webapp/WEB-INF/config/appMessages.properties
	recovery-coreextension-plugin/src/main/java/es/capgemini/pfs/core/api/persona/PersonaApi.java
	recovery-coreextension-plugin/src/main/java/es/capgemini/pfs/persona/EXTPersonaManager.java
	recovery-nuevoModeloBienesCommon-plugin/src/main/java/es/pfsgroup/plugin/recovery/nuevoModeloBienes/personas/NMBPersonaManager.java
	recovery-precontencioso-plugin/src/main/java/es/pfsgroup/plugin/precontencioso/burofax/manager/BurofaxManager.java
	recovery-precontencioso-plugin/src/web/jsp/plugin/precontencioso/burofax/grids/burofax.jsp

Former-commit-id: 59b077893afe2a80f944a8ee73a848efded9e884

+++ sql/tool/package-scripts-from-tags-list.sh
@@ -69,10 +69,18 @@ while read tagname; do
         git reset --hard HEAD
         git checkout $tagname 
         ./sql/tool/package-scripts-from-tag.sh $tagnametmp $2
+        if [ $? -ne 0 ]; then
+            echo "ERROR al generar el empaquetado"
+            exit 1
+        fi
         mkdir ./package-tags/$count
         if [ -e ./sql/tool/tmp/package/DDL/DDL-scripts.zip ];then
             cp -r ./sql/tool/tmp/package/DDL ./package-tags/$count/
             rm ./package-tags/$count/DDL/*.zip
+            for script in `find ./package-tags/$count/DDL/ -name *PREPROYECT_CNT*3.1*`;
+            do 
+                sed -e 's/SET DEFINE OFF;/SET DEFINE OFF;\nalter session set "_pred_move_around"=FALSE;\n/g' -i $script
+            done
             echo "if [ \$? != 0 ];then exit 1; fi" >> ./package-tags/run-scripts-package.sh
             echo "cd \$DIR_ORIG" >> ./package-tags/run-scripts-package.sh
             echo "cd ./$count/DDL/" >> ./package-tags/run-scripts-package.sh
@@ -92,3 +100,7 @@ while read tagname; do
 done < $1
 chmod a+x ./package-tags/*.sh
 chmod a+x ./package-tags/**/**/*.sh
+echo "----------------------------------------------"
+echo "  TERMINADO: "
+echo "     Consulta el directorio ./package-tags/"
+echo "----------------------------------------------"
